requirements.txt
numpy
Pillow
scikit-learn
pytest
